refactir(tests): Refactor test imports and fix typings
Replace barrel error imports with per-error module paths, update Tmdb
service import, and tighten Socket typing casts. Adjust watchparty test
field name and remove an unused mock import. Delete obsolete auth.utils
test file
