+++ apps/api/src/modules/auth/application/controllers/auth.controller.test.ts
@@ -1,6 +1,6 @@
 import type { NextFunction, Request, Response } from "express";
 import { describe, expect, it, vi } from "vitest";
-import { UnauthorizedError } from "../../../../shared/errors/index.js";
+import { UnauthorizedError } from "../../../../shared/errors/unauthorized-error.js";
 import { AuthController } from "./auth.controller.js";
 
 type ExecuteMock = { execute: ReturnType<typeof vi.fn> };
