+++ apps/api/src/modules/contents/infrastructure/database/repositories/tmdb-fetch-status/tmdb-fetch-status.repository.test.ts
@@ -1,11 +1,12 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
+
 import { MetadataNotFoundError } from "./errors/metadata-not-found";
+
 import { TMDBFetchStatusRepository } from "./tmdb-fetch-status.repository";
 
 const {
   dbMock,
   deleteReturning,
-  deleteWhere,
   findFirstMock,
   insertConflictUpdate,
   insertValues,
@@ -92,9 +93,7 @@ describe("TMDBFetchStatusRepository", () => {
     deleteReturning.mockResolvedValueOnce([{ id: "1" }, { id: "2" }]);
     await expect(repo.cleanupExpiredSearches()).resolves.toBe(2);
 
-    deleteWhere.mockReturnValueOnce(Promise.resolve(undefined));
     await repo.deleteDiscoverMetadata("movie");
-    deleteWhere.mockReturnValueOnce(Promise.resolve(undefined));
     await repo.deleteSearchMetadata("query");
 
     expect(dbMock.delete).toHaveBeenCalledTimes(3);
